Make `SignedWide` and `UInt::bits` visibility `pub(crate)`
We can eventually expose these as part of the public API, but I'd like
to nitpick the names first.

I'd also like to immediately cut a release in order to update the repo
location, so for now I'm removing these from the public API then we can
add them back after some discussion.