Set `resolver = "2"` and integrate proptests
Splitting the proptests out into a separate crate was largely a
workaround for the lack of `resolver = "2"`.

Now that this crate is in its own repo, we can switch to
`resolver = "2"` and use `proptest` directly.

This makes them part of the normal test workflow and thus hopefully
provides better testing ergonomics.