`Zero` trait
Extracts a `Zero` trait from the `Integer` trait which contains only the
associated `ZERO` constant and the `is_zero` method with a default
implementation.

Changes the mandatory bound for `NonZero` to be the `Zero` trait rather
than the much more restrictive `Integer`

This makes it possible to impl `Zero` on `Limb` even though it doesn't
impl the full set of `Integer` traits, and therefore use
`NonZero<Limb>`.