Impl `AsRef`/`AsMut<[LimbUInt]>` for `UInt`
Provides direct reference-based access to the inner limb array within
`UInt` as a reference to an array of `u32`/`u64` integers (depending on
the target word size).

Uses an `unsafe` pointer cast between `Limb` and its integer type,
however since `Limb` is a `repr(transparent)` newtype this is a trivial
usage of `unsafe`.