`const_assert_eq`/`const_assert_ne` fixups
These macros are designed to allow assertions around const generics,
however the previous implementation had a number of deficiencies, most
notably it only worked if the constant name was `N`.

This changes the macro "type signature" to ensure the first parameter is
an `ident`, but additionally it also contained baked-in assumptions that
its type is `usize`. This change also explicitly documents that it only
works on `usize` constants, although it'd be nice if it weren't so
constrained.