Add `BoxedUint::{from_be_slice, from_le_slice}`
Initial support for decoding `BoxedUint` from a byte slice.

Decoding is fallible and paramaterized around an expected precision.
Requiring a known precision at decode time allows the caller to ensure
that all values are the same dynamically-determined size so as to
prevent subsequent timing variability with respect to input sizes.

Decoding is currently constrained to inputs that are multiples of the
word size.