Move `modular` module to toplevel
It was previously located under `uint`.

Now that we've begun work to share the implementation between `Uint` and
`BoxedUint`, moving this module to the toplevel makes sense.

It also better maps to the public API, where it's re-exported from the
toplevel, i.e. this is not a breaking change.

Finally, less nesting makes the codebase easier to navigate.