Split `Zero` trait into `Zero`+`ZeroConstant`
The `BoxedUint` type couldn't impl the old `Zero` trait because it
requires the value be bound to a constant instead of computed at
runtime.

This commit splits the original trait into two traits, adding a
`fn zero()` to `Zero` which can be used for `BoxedUint`.

A blanket impl links the two.

This makes it possible to add a `Zero` trait impl for `BoxedUint`, which
is included in this commit.