Make `BoxedUint::rem_vartime` infallible using `NonZero`
Now that #335 has been merged allowing `NonZero<BoxedUint>` to be
expressed, we can use it as the operand for remainder, eliminating the
need to handle division by zero at the type system level.

This dramatically simplifies the implementation of
`BoxedResidueParams::new` since now all of the remainder calculations
are infallible, and is more consistent with the other remainder
functions.