Assert `Boxed*` size equivalance using `bits_precision`
Several functions on the `Boxed*` types don't yet support implicit
widening (#312) and will panic if two or more operands are not the same
size (NOTE: we should eventually fix this)

Some of these functions previously had debug asserts that the number of
limbs are equal, however that's less helpful information when trying to
debug these problems than the precision in bits.

This changes all of the assertions for size equality to use
`bits_precision`.