BoxedUint: share `Mul` impl with `Uint`
Uses a common implementation of schoolbook multiplication factored into
the `uint::mul` module.

For now, it's impl'd as a macro to share a common implementation between
`const fn` and heap-backed `BoxedUint`, since `const_mut_refs` is not
yet stable and needed to be able to write into a non-array output.