Eliminate some timing variability in `montgomery_mul`
Always performs `sub_vv` on the lower half of the output, and
conditionally assigns `upper` to `lower` using `subtle`.

It doesn't matter if we unconditionally subtract, since the output will
be overwritten in the case where it's "unnecessary" (but actually
necessary for constant-time operation).

There are still some data-dependent branches, however, which have been
annotated with a TODO.