Impl `Wrapping*` traits from `num-traits`
Until now we have used traits defined in this crate, primarily because
many of these traits are predicates that return `bool` and in
constant-time code we want those to return `Choice`/`CtOption`.

`Wrapping*` is an example of the sort of trait we'd want to incorporate
unchanged, and there are potentially others, so this adds `num-traits`
as a hard dependency (previously our only hard dependency was `subtle`).
Note that `num-traits` itself has no transitive dependencies (or rather,
they're optional but not enabled).

The `Wrapping*` traits have bounds on the corresponding op traits being
impl'd with `Self` operands e.g. `WrappingAdd: Add<Self, Output = Self>`
so this PR also adds impls of those traits.

We've previously avoided these as in `std` they panic on
overflow/underflow in debug builds and silently wrap in release builds.
This PR always panics.

This required some changes to the `Mul` impls which were conditional on
`ConcatMixed` and implicitly widened. To accomodate impls which are
always available and require no bounds (in order to allow us to impl
`WideningMul`), and renames the following:

- `Uint::mul` -> `Uint::widening_mul`
- `Uint::mul_wide` -> `Uint::widening_mul_split`