Add `Zero::set_zero` method
Adds a method to the `Zero` trait for setting its value to zero
in-place, subsuming the previous `BoxedUint::set_to_zero` function.

It's inspired by a similar method on `num_traits::Zero` (which we don't
use so `is_zero` can return `Choice` and use `ConstantTimeEq` to compare
values in constant-time).

https://docs.rs/num-traits/latest/num_traits/identities/trait.Zero.html#method.set_zero