Bernstein-Yang cleanups; MSRV 1.73 (#437)
Refactors the `CInt` type into `Uint62L` which always uses a 62-bit
unsaturated representation. This is mostly to reduce the overall number
of const generic parameters involved, since Bernstein-Yang always uses a
62-bit unsaturated representation (at least as currently implemented).

Moves the `S` parameter from `invert` into the `BernsteinYangInverter`
struct itself. This allows the type to always "know" its associated
inverter, rather than leaving it up to the end user to calculate.

Also changes the `invert` method to accept a `Uint` as input.

Bumps MSRV to 1.73 since it uses the `int_roundings` feature to compute
the number of unsaturated limbs from the number of saturated limbs.