Add `Inverter` trait
Adds a trait for obtaining a `BernsteinYangInverter` for the given
`Uint` modulus, precomputing the number of unsaturated limbs needed for
the 62-bit unsaturated form used by Bernstein-Yang at compile time, and
associating the proper `BernsteinYangInverter` for a given `Uint` size
as an associated type.

This should make it possible to obtain a `BernsteinYangInverter` for a
`Residue`/`DynResidue`, though that's left as a followup exercise.