Use full Montgomery mul for `BoxedResidue::{mul, square}`
Previously "almost Montgomery multiplications" were used which return a
result which is only reduced modulo the bit size of the modulus, but may
still equal or exceed the modulus.

This commit goes back to using a multiply followed by a full Montgomery
reduction, and adds more asserts that the result is fully reduced.
This incurs a performance hit, but AMM is likely "unsafe" to use outside
of this crate unless we decide to fully encapsulate the inner
representation of `*Residue` types.

Note that this does not ensure the output of modpow is fully reduced,
which is left to be fixed in a followup.

This commit includes a regression test from the issue where this was
reported (#441) and additionally adds some general proptests for
`BoxedResidue::square`.