Add new `Inverter` trait; rename `PrecomputeInverter`
Adds a trait for abstracting over various sized `BernsteinYangInverter`
(and prospectively a `BoxedBernsteinYangInverter`) types, which also
opens the door to impl'ing `PrecomputeInverter` for `*ResidueParams`,
where the inverter would accept a `*Residue` as input/output and invert
in Montgomery form.

The `PrecomputeInverter::Inverter` type is now bounded on the new
`Inverter` trait, making it useful in generic code.