Bernstein-Yang: fallible constructors; return `CtOption`
Adds a `ConstChoice` to the `const fn` versions of a Bernstein-Yang
constructor as well as computing inversions on non-`NonZero` inputs.

Changes the `Inverter` trait to return `CtChoice` on non-`NonZero`
inputs (`NonZero` inputs are still infallible).

The `PrecomputeInverter` impl on `Uint` now panics if called on an even
modulus, which may not be ideal, but allows a prospective impl on
`*ResidueParams` to be infallible, which is the main way this is
intended to be used (though the impl on `Uint` will be the core).