Extract sealed `PrecomputeInverterWithAdjuster` trait (#449)
This trait is needed for the Montgomery form types to instantiate a
Bernstein-Yang inverter which operates in the Montgomery domain, by
passing R^2 as the adjuster.

It's sealed for now, but could be made public if needed (which might be
the case for writing generic trait bounds).

It's extracted from the `PrecomputeInverter` trait, where previously the
impl for `DynResidueParams` had a `todo!()` about what to do with this
method. Splitting it out into a separate trait resolves this issue.