BoxedResidue: add a second conditional subtraction to `pow`
Proptest found an edge case (checked in) where the output of
`BoxedResidue::pow` was still greater than the modulus.

The original implementation in `num-bigint` this was adapted from
conditionally subtracted (not in constant-time, mind you) and then
checked again, conditionally applying a full remainder in the event it
still exceeded the modulus.

We might need to end up doing something similar, but for now,
subtracting the modulus again seems to at least address the proptest
failure.