Uint: change `split_mixed` to accept `self`; make `pub`
The `split_mixed` method provides a `const fn`-friendly way to split
numbers.

This refactors it to be a method that accepts `self` as an argument so
it can be called as `uint.split_mixed()` and also exposes it as `pub`,
allowing downstream users to use it in `const fn` contexts.

A newly added `SplitMixed` bound typechecks the sizes of the outputs to
ensure that they're a valid combination, which is the best we can do
absent something like `generic_const_exprs`.