Fix `bernstein_yang_nlimbs!` calculation
The previous calculation of the number of unsaturated 62-bit limbs
needed to represent an integer of a given size was incorrect, leading to
miscomputed results as seen in #606.

This commit switches to a much simpler calculation based on
`div_ceiling(62)`, and also adds a const assertion that the computed
number of limbs is sufficient to hold a `$bits`-sized integer.