Initial constant-time stack-allocated Bernstein-Yang
The previous implementation runs in variable-time with respect to `g`.
However in the event both inputs are secret a fully constant-time
implementation is required.

This implements the method described in section 11 of
https://eprint.iacr.org/2019/266.pdf and more specifically this Python
code from Figure 11.1:

    from divsteps2 import divsteps2

    def iterations(d):
      return (49*d+80)//17 if d<46 else (49*d+57)//17

    def gcd2(f,g):
      assert f & 1
      d = max(f.nbits(),g.nbits())
      m = iterations(d)
      delta, fm, gm, P = divsteps2 (m,m+d,1,f,g)
      return abs(fm)

    def recip2(f,g):
      assert f & 1
      d = max(f.nbits(),g.nbits())
      m = iterations(d)
      precomp = Integers(f)((f+1)/2)^(m−1)
      delta, fm, gm, P = divsteps2(m,m+1,1,f,g)
      V = sign(fm)*ZZ(P[0][1]*2^(m-1))
      return ZZ(V*precomp)

Instead of bounding the loop on `g` reaching zero, this instead computes
a fixed number of iterations relative to the highest bit of either `f`
or `g` after which the algorithm will converge, then runs for that
number of iterations instead.

This results in about a 22X performance impact:

    Bernstein-Yang invert, U256
            time:   [36.934 µs 37.151 µs 37.391 µs]
            change: [+2256.7% +2267.7% +2278.4%] (p = 0.00 < 0.05)
            Performance has regressed.

The previous implementation which is variable-time with respect to `g`
is preserved as well, for now as `gcd_vartime`, but it would also be
nice to add an `inv_mod_vartime` as well.