Constant-time heap-allocated Bernstein-Yang
The previous implementation runs in variable-time with respect to `g`.
However in the event both inputs are secret a fully constant-time
implementation is required.

This implements Bernstein-Yang in constant-time with respect to both
parameters by computing a worst case number of iterations for the
algorithm to converge, partially sharing the implementation with #632.