Add `UintRef`
Adds a reference newtype for a `Limb` slice which can be used as an
abstraction for implementing shared functionality for both `Uint` and
`BoxedUint`.

The initial implementation supports a limited amount of functionality
but supports usage as both `&UintRef` and `&mut UintRef`.

This includes a common implementation of `fmt` impls shared between
`Uint` and `BoxedUint`, replacing a copy-paste implementation.

Closes #756