DRY out Monty params between const and non-const forms
Creates a `modular::params` module containing a set of `const fn`
functions for computing Montgomery params, and uses these functions for
both the `const_monty_params!` macro as well as `MontyParams`.

Ideally we could just use `MontyParams` for this (as it has a `const fn`
constructor) and eventually this can probably be factored there (i.e.
like it was before this PR). However, doing that properly would require
`adt_const_params` (rust-lang/rust#95174), which is why we have
`const_monty_params!` in the first place, so for now this is the best we
can do.