Revert "Rename `div_by_2` to `shr1` (#882)"
This reverts commit 5f48318a49b72500af825d1ef0d70c32a3319fc9.

My original rationale for this was to share code with a larger `shr`
implementation. That could be done with repeat applications of `shr1`,
though using a Montgomery multiplication would probably be faster.

However, I wound up not needing those, which perhaps makes this change
superfluous. Since I noted it's six of one, half dozen of another, it
would probably be better to just use the original name in that case.