Add `Reduce` trait for fixed modulus reduction. (#887)
This trait comes from `elliptic_curve::Reduce`.

I thought this would fit better here as `cyrpto-bigint` already hosts a
lot of `modular` arithmetic functionality.

Note that it slightly differs from `elliptic_curve::Reduce` because the
input type is not bound by `Integer`, and there is no associated `Bytes`
(and no `reduce_bytes` function). That is because `Reduce` can be
implemented both on the input type and on `Bytes` as separate
implementations, if the input type is not bounded by `Integer`.