Clean up GCD API (#923)
This mostly avoids exposing the separate `safegcd`/`bingcd`
implementations. Both use the same pre-processing for inputs so the
active variant only needs to be selected in `Odd<Uint>`. The
`safegcd(_vartime)` and `bingcd(_vartime)` methods are only exposed for
benchmarking at the moment.

Because a `const gcd` method on each integer type would have a higher
priority than the trait method, it would produce type errors when trying
to compute the GCD between different types (like `Uint` and `Int`). I've
gone with the name `gcd_uint` for these methods.

---------

Signed-off-by: Andrew Whitehead <cywolf@gmail.com>