Impl `(Fixed)Integer` and `Signed` for `Int`
Followup to #942

This factors the parts of `Integer` that `Int` currently doesn't
implement onto `Unsigned`, along with things that definitely belong on
`Unsigned`.

In particular, there's a TODO for `WrappingMul`, where it appears `Int`
lacks an implementation.

The two types diverge on `Div`-by-`NonZero`, where `Int` currently
returns a `CtOption`. From what I can tell this is an artifact of the
implementation (the potential failing case is the quotient when computed
as a `Uint` exceeding the size of an `Int` from what I can tell). As it
were, the `DivAssign` impl uses `expect` in such a case. Ideally, the
two could be unified and have common, `CtOption`-free bounds on
`Integer` instead.

There are several places `From` impls and methods of `Integer` expect
it's infallibly possible to create an `Integer` from a `Limb`, or a
`u32`, or a `u64`. But this isn't possible with `Int<1>` on 32-bit
platforms, where a `u32` or anything bigger (which includes `Limb`)
could overflow the maximum positive value.