Add `One` trait (like `Zero`)
Adds a trait for the multiplicitive identity element of `Self`, similar
to `num_traits::One` but with support for constant-time comparisons and
predicate methods like `is_one`.

It follows the same basic shape as the existing `Zero` trait.

The `Integer` trait is now bounded on it, and this removes the previous
`one` and `one_like` methods from `Integer`, which are now available via
the new trait.

`One` is impl'd for all the same types as `Zero`, and notably it can
also be impl'd for `NonZero<T>` and `Odd<T>` where `Zero` definitionally
cannot.