Optimize checked multiplication and squaring (#983)
This optimizes checked (and saturated) multiplication and squaring to
perform a wrapping multiplication with a carry, and check for overflow
by a more efficient method. For boxed integers this means the allocation
is no larger than needed. Performance is 30+% improved at the 4096 bit
size and more at larger sizes.

A few missing methods are added for consistency: inherent `checked_mul`
methods, as well as `saturating_mul` and `saturating_square` for
`BoxedUint`.

`Int` could also use some improvements here but I think that's a
separate PR.

Resolves #810

---------

Signed-off-by: Andrew Whitehead <cywolf@gmail.com>