Make AMM-based modular `pow` function generic
The AMM modpow implementation was previously specialized to
`BoxedMontyForm`. This extracts the implementation into the toplevel
`pow` module and makes it generic.

Note it doesn't yet add support for using it with `MontyForm`, but that
should be a straightforward next step after this.

Making this work involved defining a `MontyMultiplier` extension trait
which I've called `AmmMultiplier`, however it's marked `pub(crate)` and
is otherwise entirely internal.

Constructing `BoxedMontyForm` from an AMM now has an internal helper
function called `BoxedMontyForm::from_amm`, both to permit reuse and
because the methods needed to implement it with optimum performance
aren't defined on the `Integer`/`Unsigned` traits.

This has no impact on benchmarks:

Boxed Montgomery arithmetic/modpow, BoxedUint^BoxedUint
                        time:   [30.203 ms 30.263 ms 30.325 ms]
                        change: [−0.1536% +0.2861% +0.6654%] (p = 0.17 > 0.05)
                        No change in performance detected.