Add native linux- and darwin-aarch64 tests (#1027)
Removes the `test-cross` target for `aarch64-unknown-linux-gnu`,
preferring to run this test against GitHub’s native Linux arm runners.
Should give a considerable speedup from the combination of not needing
to be virtualized and not needing to set up the cross environment.

Also adds a test case against Darwin on `macos-latest` to get another
aarch64 platform, and avoids running an empty `deps` action if there is
none.