Rename `random_mod` ↦ `random_mod_vartime`
Rejection sampling means that `random_mod` is inherently variable-time,
and the current name violates the contract in the `README`:

> All functions contained in the crate are designed to execute in
> constant time unless explicitly specified otherwise (via a `*_vartime`
> name suffix).

This renames `random_mod` to `random_mod_vartime`, adding in deprecated
aliases at the old names for backwards-compatibility. It preserves the
`ConstantTimeLess` functionality for good measure and to not needlessly
increase the amount of timing information leaked.

Of course, `random_mod` leaks a pretty small amount of timing
information; by my reckoning, at worst an average 2 bits per call, and
approaching 0 as the modulus approaches `2^n-1`. (We would have exactly
0 at `2^n` if we did a check for powers of 2, but this is currently a
worst case for our `random_mod`; if the user knows they have a power of
2, they should use `random_bits` instead.) So this is a bit of an edge
case; but I argue that the `vartime` naming here reduces surprise.