Replace `ConstCtOption` with `ctutils::CtOption`
Analogous to what #1035 did for `ConstChoice`, this extracts the
necessary methods onto `ctutils::CtOption` then replaces the type
wholesale using the implementation in `ctutils`.

There were a few tricky parts, namely we had type-specific
implementations of `expect` and `unwrap_or`.

It is actually possible to write a generic `const fn` implementation of
`CtOption::expect`, avoiding the problems with destructors and the
unstable `Destruct` trait, by bounding on `T: Copy` and returning a copy
of the inner value. This is implemented as `CtOption::expect_copied` so
we can still support `expect` for non-`Copy` types.

Unfortunately there's no analogous generic `const fn` solution for
`CtOption::unwrap_or`, as this needs a constant-time selector function
and even function pointers can't be used from a `const fn`. So this does
the only thing we can do to replace it with some level of abstraction:
a `ctutils::unwrap_or!` macro.

See also: RustCrypto/utils#1274 which extracts some needed methods onto
`ctutils::CtOption`.