Replace `ConstantTimeSelect` with `ctutils::CtSelect`
The `ConstantTimeSelect` trait was a workaround for having a trait-based
API that works for `Boxed*` and non-boxed types, since we can't impl
`subtle::ConditionallySelectable` on `Boxed*` types since it has a
`Copy` bound.

`ctutils::CtSelect` accomplishes the same goals by having only a `Sized`
bound (since we need to be able to return selected values).

This commit completely replaces the `ConstantTimeSelect` trait with
`ctutils::CtSelect`.