word: cleanup helper functions
Cleans up the internal helper functions in the `word` module, adding
a `choice_from_*` prefix to functions that return a `Choice`, and
getting rid of `*word*` in the function name since it's implicit from
the `word::*` prefix which is used ubiquitously when calling these
functions.

Also changes the argument order of `select` and `select_wide` (as newly
renamed) so the `Choice` is the last argument, which is more consistent
with the rest of the codebase.