Use `ConstCtOption` for checked operations (#1052)
Uses it both as the return type for ops described by `Checked*` traits
as well as the inner type for the `Checked` wrapper.

This seemingly uncovered some issues with the division implementation,
including the possibility that checked division was somehow
short-circuited when using the previous implementation of the `NonZero`
type, as with this change it's now invoking a function that panics with
a zero divisor, when instead it still needs to always call the same
function and get a result without panicking, even if the divisor is
zero.

To solve the general problem, `NonZero::new` now also bounds on `One`
and `CtSelect` (in addition to `Zero`), and automatically uses
`One::one_like` as a placeholder value in the `ConstCtOption` it returns
in the event the provided value is equal to zero. That way we never
construct an invalid `NonZero` containing an inner `0` value.