Convert remaining `subtle` usages to `ctutils`
Except for impls of `subtle` traits (which we can make optional), this
gets rid of all of the remaining usage of `subtle` in the internals of
the library, replacing it with `ctutils`.

After this, we can make `subtle` optional, and rename `ConstChoice` to
`Choice` and `ConstCtOption` to `CtOption`, removing the split between
the two types.

This commit fills in various trait impls as needed, and also adds `CtEq`
and `CtSelect` bounds to the `Monty` trait.