Make `subtle` an optional dependency
After a long string of PRs to convert all previously `subtle`-based
constant-time functionality to `ctutils`, #1054 removed the last of the
hard dependencies on `subtle`, and we can now make it optional.

Enabling the feature provides legacy `subtle` trait impls, and also
enables the corresponding `subtle` feature on `ctutils`, and if it's
enabled, also on `hybrid-array`.