Get rid of `Const*` prefix on `Choice` and `CtOption`
At long last, the split is healed and we have one `Choice` and one
`CtOption` type that can both do it all.

The implementation of both actually contains large parts of the old
`ConstChoice` and `ConstCtOption` types extracted into the `ctutils`
crate, along with more `subtle`-like functionality.

This commit renames all of the following usages:
- `ConstChoice` => `Choice`
- `ConstCtOption` => `CtOption`

For legacy compatibility, deprecated type aliases are made available for
the old names, pointing to the new ones.