Add `From` impls between `ByteArray` and `EncodedUint` (#1064)
One thing lost from #1016 is the ability to convert from `Uint::Repr`
and `hybrid_array::Array<u8, _>`.

Arguably these cases should use the dedicated
`ArrayEncoding`/`ArrayDecoding` traits for this purpose instead, but for
the sake of retrofitting the changes from #1016 this adds back the
ability to do these conversions for `EncodedUint` up to 16 limbs.